Add a release buffer in the store to keep queries retained longer
Summary:
This commit adds a "release buffer" to the store to keep a specific (configurable) number of queries temporarily retained even after they have been released by the caller.

=== motivation
The motivation for this change is to allow data not to be as aggressively garbage collected by Relay, and increase the likelihood of data in the store to be able to be re-used during the app lifetime (e.g. when navigating from a list view to a detail view, and then back to the same detail view). This is particularly important since we now support the capability of reusing data in the store with the fetch policies in Relay Hooks, but which wouldn't be very useful in practice without this change, or without completely disabling garbage collection.

Longer term, this will also allow us to completely remove the network response cache (at the newtork layer) in favor of reusing data directly in the store; the latter approach has a few advantages, like always reusing up to date data / being resilient to updates (the network response cache needs to be cleared after any mutation or subscription), as well as not necessarily being constrained by an arbitrary cache size.

=== this commit

Specifically, this commit adds a buffer to the relay store which has a configurable capacity.

The way it works is the following:

- Operations are retained as normal.
- When an operation is released by the original owner, instead of immediately releasing it and scheduling a GC run, the operation will be moved to the release buffer.
- If the release buffer reaches capacity when a new operation is added, the least recent operation will be released, and a GC run will be scheduled.

So for example, with a buffer capacity of 10, this will have almost the same behavior as a network response cache of size 10, with the added benefit that the cache won't be cleared every time an update occurs.

Note that providing a buffer capacity of 0 is equivalent to current behavior.

Reviewed By: josephsavona

Differential Revision: D17789365

fbshipit-source-id: d93d13792a2c121b454cfd773294ca2eedc51abf