speed up getFragmentIdentifier
Summary: We already have a pre-computed identifier for the fragment owner + variables, this uses that to avoid calling `stableCopy` on the larger object and just call on the fragment variables. I **think** even the fragment variables might be sorted by the compiler or we could sort them earlier to avoid sorting on every fragment read.

Reviewed By: jstejada

Differential Revision: D17799185

fbshipit-source-id: 8c684bd39836c2969f6bab375d3a8511dca04326