merge CompilerValidations into transforms
Summary:
The transforms already do a lot of validations at various stages of the transform and can be configured differently if needed.

The only thing that was different for validations was that they returned `void`. By changing them to identity functions that might throw errors, we can eliminate one concept from the compiler.

Reviewed By: alunyov

Differential Revision: D17906715

fbshipit-source-id: ba34ee371a49ae3c4c692daacea43ae7e5f8f2c1