sort printed documents by name
Summary:
The fragments were ordered by insertion order which can sometimes cause a small change to trigger a complete re-ordering of the generated files and a diff that is a lot larger than it ought to be.

Sorting the printed document (roots first, then split operations, finally fragments) should help get us a more stable output.

Reviewed By: josephsavona

Differential Revision: D17988183

fbshipit-source-id: 37673d5876c39f2bb3d968b1903bc2df33ef0d78