Fix resolving fragment snapshot in refetch/pag containers when owner variables changed
Summary:
In D17763049 and D17768223 (which landed as D17803860), we stopped using variables from context in the refetch and pagination containers, which meant that we stopped checking if the variables had changed.

This was fine for the most part, since now the variables are propagated through the owners in the fragment reference, so when we `setProps` on the FragmentResolver, it will use the updated variables.

However, unlike the Fragment Container, the pagination and refetch contaienrs only ***conditionally*** update the resolver with the next props; specifically **only if we havent made a refetch or pagination request***. This made it so that if a container happened to have refetched or paginated, if it received new props with new fragment references that contained new owner variables, it would not update them in the resolver:

https://our.intern.facebook.com/intern/diffusion/FBS/browse/master/xplat/js/RKJSModules/Libraries/Relay/oss/react-relay/ReactRelayRefetchContainer.js?commit=13d1dd70ff527f279244b3c992aea3c4bbd21276&lines=182-184

I'm not sure why this condition exists in the first place, and why we don't unconditionally always `resolver.setProps(nextProps)`, like the Fragment Contaiener does; the history on this decision is a bit fuzzy.

To fix this, this commit restores the check to check if variables from the fragment owners have changed, and then re-create a new resolver then. I decided /not/ to unconditionally update the resolver with the new props since I don't know the original motivation behind that decision and am afraid it might break something else.

Reviewed By: alunyov

Differential Revision: D18014700

fbshipit-source-id: d0a9ccf8a5afe2a7de7e543bcdf2fb756a9a37c2