preloadQuery(): allow passing a full query, add docs
Summary: Allows passing a raw query to `preloadQuery()`, which makes it usable without the special $Parameters file - ie makes it easily usable in OSS today. This seems like a pretty reasonable change. Yes, it's typically more optimal to avoid eagerly downloading a full query AST. But if you do already have one, there's no reason not to be able to call this API.

Reviewed By: jstejada

Differential Revision: D18110449

fbshipit-source-id: eb7ad65e644b486b61491c06e6ad9c072e8d7a00