Data Invalidation: Update check() result to reflect when individual records are stale
Summary:
This commit updates the logic for `DataChecker.check()` to account for individually invalidated records.

Specifically, if `DataChecker` encounters a stale record during the check, it will short circuit the check and return `stale` as the result.

Additionally, the integration tests added here detected an issue with invalidating newly created records, which this commit fixes

Reviewed By: josephsavona

Differential Revision: D18690069

fbshipit-source-id: 751a1c1825a725a4b6fe0942433946c7768fe039