Data Invalidation: Primitives for subscribing and checking invalidation state
Summary:
This commit adds 3 new primitives to the Relay store which allow to observe the invalidation state of specific data ids in the store.

Our current invalidation primitives serve to invalidate data, and data that is marked as stale will /only/ be refetched the ***next time*** a query is evaluated (e.g. upon navigation to a new page), but we need a way to trigger immediate re-evaluation, which is what these primitives are for. Specifically, the purpose of these new primitives is to be able to handle the following use cases:

* Refetching data that is already visible in the current page. Since no navigation is ocurring, we won’t re-revaluate the queries for the current page, so even if some records are stale we won’t refetch the query since we are not checking the query again.
* Refetching data when navigating to a previous view that was never unmounted; since the view wasn't unmounted when navigating back, the queries for the route wont be re-evaluated, meaning that even if some records are stale we won’t refetch the query since we are not checking the queries again.

In the next diff, we will add a React hook to be able to subscribe to invaldiation state from product code.

Reviewed By: josephsavona

Differential Revision: D18908948

fbshipit-source-id: 5a729e9d49a588a96c1fa77a4592ee595ed60f7f