Allow language plugin to specify how modules should be named. (#2866)
Summary:
This allows the language plugin to specify how modules should be named.

Context: In Reason, `.` is not allowed in file names except for with the file extension. This means that the language plugin I'm working on for Reason needs a way to have modules/generated files be named `OperationName_graphql.re` instead of `OperationName.graphql.js` in order for it to work.

Please let me know what you think!
Pull Request resolved: https://github.com/facebook/relay/pull/2866

Reviewed By: jstejada

Differential Revision: D19815165

Pulled By: alunyov

fbshipit-source-id: 5291023ea72fc276e6f958aa1ef7a42f10b4ee12