transforms benchmarks
Summary:
Adds a criterion benchmark that runs all of our (current) transforms against all of the documents from across all the compileRelayArtifacts fixtures. I accumulated all the queries/fragments from those fixtures into a single .graphql file, prefixing the definitions from each fixture with the fixture name (and updating fragment spreads accordingly).

This means we can run `cargo bench --bench transforms` and get benchmark output, which is saved and compared to the previous run. Unfortunately on a MBP it's a bit noisy, i see variance of 10% run to run, so we'll have to figure out a better setup for stable benchmarking. Would be helpful if we could get instruction counts.

Reviewed By: kassens

Differential Revision: D19862306

fbshipit-source-id: 54a4a3f3bd31bb68d101f4833de9142082959e81