remove base fragments from output
Summary:
This adds a small transform that removes base fragments from a `Program`. We apply this on the reader program, which is used to determine what artifacts are written for fragments.

In the JS compiler, I realized we do this step *after* all the transforms, but that's some wasted work in transforming these reader nodes for base fragments again just to throw them out.

Reviewed By: tyao1

Differential Revision: D20170787

fbshipit-source-id: d4bcc885451c51bcff0225753827b1f279220711