transform functions can return any selection
Summary:
This diff switches transform functions for `Selection` variants to return the `Selection` enum instead of the variant.

This allows those functions to change the type of selection with the cost of a bit more verbosity in each transform that now has to return the enum variant.

I briefly considered implementing `Into<Selection> for LinkedField`, so you could write

  LinkedField { ... }.into()

instead of

  Selection::LinkedField(Arc::new(LinkedField { ... }))

but I was concerned that `into()` hides a bit too much whats going on? We could experiment with that on top of this diff.

This diff combines pretty well with D20368755 (addition of `Transformed::ReplaceMultiple`) to make inlining of selections really easy as seen in the `apply_fragment_arguments` transform stacked on top of that.

Reviewed By: josephsavona

Differential Revision: D20390240

fbshipit-source-id: d4b13e394b2e6787774ca26fbe6b11fb319f6373