rework persisted query re-using
Summary:
Previously, the flow to avoid re-persisting documents we already persisted before worked like this:
1. Generate flow, query test, AST
2. Hash those
3. See if the existing generated file has the same hash.  If that's the case, just mark the file as up to date.

There are various changes that might affect one of these inputs in the hash, but not he others, so to guarantee an up to date file, we need to hash all those inputs. In fact, this isn't even enough! For example, if the formatting of the generated file changes, we would need to force-rebuild or bump the cache breaker.

We can also see that we generate all the inputs anyway.

The problem is that duplicating this hashing in Rust would be difficult and thining about it, we can actually improve this to just:

1. Generate all inputs as before.
2. Just hash the document text
3. If the existing hash matches, we re-use just the persist id instead of calling the persist script.
4. We always check the contents against disk after we insert the persist id.

Note that only hashing the query text is save as the query id is the only part we're re-using and that only depends on the query id and doens't care about the type text or AST.

Reviewed By: alunyov

Differential Revision: D20818398

fbshipit-source-id: cb74eec08ad1e64dd738527b502ea1b503174e0f